requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.2.0
chardet==3.0.4
discord.py==1.5.0
idna==2.10
multidict==4.7.6
Pillow==8.0.0
yarl==1.5.1